reverting, implementing portal placement viz in separate component to avoid ref mount issues
