+++ .gitignore
@@ -23,3 +23,6 @@ dist-ssr
 *.sln
 *.sw?
 .vercel
+
+# Docs
+INSTRUCTIONS.md
